fix(main): Docs for FastAPI sub-apps visible
The proxy's openapi_url, docs_url and redoc_url used to 'overwrite' the docs of the url that's being served, and once that was accessed, it would still refer to the openapi_url of the proxy. By disabling these three, docs urls for FastAPI apps being used with this proxy will function properly.